\name{Inside}
\alias{Inside}
\title{Inside}
\usage{
Inside(value, Start, End)
}
\description{
Returns a bollean value or vector whether the value of the value vector belongs to a given segment [Start, End].
The function also works if the Start and the End parameters are the vectors of the same size.
}
\examples{
Inside(4, 3, 5)
[1] TRUE
}
